Conflicts:
	engine/report/sc_chart.cpp
	engine/report/sc_report_html_sim.cpp

+++ engine/report/sc_report_text.cpp
@@ -570,16 +570,16 @@ void print_text_performance( FILE* file, sim_t* sim )
     date_str = date_str.substr( 0, date_str.size() - 1 );
   util::fprintf( file,
                  "\nBaseline Performance:\n"
-		 "  RNG Engine    = %s\n"
-		 "  Iterations    = %d\n"
-                 "  TotalEvents   = %" PRIu64 "\n"
-                 "  MaxEventQueue = %" PRIu64 "\n"
+                 "  RNG Engine    = %s\n"
+                 "  Iterations    = %d\n"
+                 "  TotalEvents   = %lu\n"
+                 "  MaxEventQueue = %lu\n"
                  "  TargetHealth  = %.0f\n"
                  "  SimSeconds    = %.0f\n"
                  "  CpuSeconds    = %.3f\n"
                  "  WallSeconds   = %.3f\n"
                  "  SpeedUp       = %.0f\n"
-                 "  EndTime       = %s (%" PRIu64 ")\n\n",
+                 "  EndTime       = %s (%.0f)\n\n",
 		 sim -> rng().name(), sim -> iterations,
                  sim -> event_mgr.total_events_processed,
                  sim -> event_mgr.max_events_remaining,
@@ -588,7 +588,8 @@ void print_text_performance( FILE* file, sim_t* sim )
                  sim -> elapsed_cpu,
                  sim -> elapsed_time,
                  sim -> iterations * sim -> simulation_length.mean() / sim -> elapsed_cpu,
-                 date_str.c_str(), cur_time );
+                 date_str.c_str(),
+                 static_cast<double>( cur_time ) );
 }
 
 // print_text_scale_factors =================================================
